refactor: 타이머가 줄어들때 전체 리렌더링이 아닌 타이머 부분만 리렌더링 되도록 리팩토링(#214)
