+++ src/hooks/useTryAnswer.ts
@@ -6,7 +6,6 @@ import { extractRemainingChances } from '#/shared/utils'
 import { extractColorStyle } from '#/shared/utils/extractColor'
 import { extractFontStyle } from '#/shared/utils/extractFontStyle'
 import { extractPatternStyle } from '#/shared/utils/extractPattern'
-import { TryAnswerResponse } from '#/types/tryAnswer'
 import { useEffect, useMemo, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { useParams } from 'react-router'
@@ -21,9 +20,8 @@ const useTryAnswer = () => {
   const maxChances = 3
   const [chances, setChances] = useState<number>(maxChances)
   const [isShaking, setIsShaking] = useState(false)
-  const [timeLeft, setTimeLeft] = useState<number | null>(null)
   const [isFlipped, setIsFlipped] = useState(false)
-  const [response, setResponse] = useState<TryAnswerResponse | null>(null)
+  const [remainingSeconds, setRemainingSeconds] = useState<number | null>(null)
 
   const { data: letterData, isError: isLetterError } = useGetLetterData({
     uuid: uuid!,
@@ -63,44 +61,25 @@ const useTryAnswer = () => {
           setChances(3 - answerStatusData.data.try)
         } else {
           setChances(0)
+          setRemainingSeconds(answerStatusData.data.remaining_seconds)
         }
       }
     }
   }, [answerStatusData])
 
-  useEffect(() => {
-    if (
-      (chances === 0 && answerStatusData.data.remaining_seconds) ||
-      (chances === 0 && response?.remaining_seconds)
-    ) {
-      const remainingSeconds =
-        response?.remaining_seconds || answerStatusData.data.remaining_seconds
-      setTimeLeft(remainingSeconds)
-      const timer = setInterval(() => {
-        setTimeLeft((prev) => {
-          if (prev && prev > 0) return prev - 1
-          clearInterval(timer)
-          setChances(maxChances)
-          // refetchAnswerStatus()
-          return null
-        })
-      }, 1000)
-      return () => clearInterval(timer)
-    }
-  }, [chances, answerStatusData, response, refetchAnswerStatus])
-
-  const handleWrongAttempt = (remaining_seconds: number) => {
+  const handleWrongAttempt = () => {
     if (chances > 0) {
       setChances((prevChances) => {
         const newChances = prevChances - 1
-        if (newChances === 0) {
-          setTimeLeft(remaining_seconds)
-        }
         return newChances
       })
       setIsShaking(true)
       setTimeout(() => setIsShaking(false), 500)
     }
+  }
+
+  const resetChances = () => {
+    setChances(maxChances)
   }
 
   const tryAnswer = async (inputValue: string) => {
@@ -114,16 +93,14 @@ const useTryAnswer = () => {
         return
       }
 
-      setResponse(response)
-
       if (response.success) {
         setIsCorrect(true)
         setResponseMessage(t('tryAnswer.correctAnswer'))
         setButtonText(t('tryAnswer.checkAnswer'))
         setIsFlipped(true)
       } else {
-        const remainingSeconds = response.remaining_seconds ?? 0
-        handleWrongAttempt(remainingSeconds)
+        setRemainingSeconds(response.remaining_seconds ?? 0)
+        handleWrongAttempt()
         if (response.hints && response.hints.length > 0) {
           refetchAnswerStatus()
         }
@@ -133,7 +110,6 @@ const useTryAnswer = () => {
       }
     } catch (error) {
       throw error
-      setResponseMessage(t('error.unexpectedError'))
     }
   }
 
@@ -170,7 +146,6 @@ const useTryAnswer = () => {
     buttonText,
     chances,
     isShaking,
-    timeLeft,
     tryAnswer,
     isFlipped,
     letterData,
@@ -181,6 +156,8 @@ const useTryAnswer = () => {
     cycle,
     hints,
     isNotFound,
+    remainingSeconds,
+    resetChances,
   }
 }
